Store cooked per-frame model bounds as i16
LevelModelFrameBoundsRecord holds one conservative sphere per cooked
animation frame. Cortex Tech Demo 0.4b cooks 1,914 of them and every
value lies in -52..109, yet each record took four i32s (30,624 B of
.rodata). Narrowing to i16 halves the table (-15,312 B) and the one
consumer widens back to i32 before any arithmetic. The cooker now
refuses a package whose bounds leave the i16 range instead of
truncating them.

+++ engine/crates/psx-level/src/lib.rs
@@ -1953,12 +1953,16 @@ pub struct LevelModelClipBoundsRecord {
 /// before instance yaw and world translation. `radius` is in engine
 /// world units and intentionally conservative so culling never clips
 /// a visible animated limb.
+///
+/// Stored as `i16` because the table has one record per cooked frame
+/// (1,914 in Cortex Tech Demo 0.4b, 30.6 KB as `i32`) and every value is a
+/// model-sized distance; the cooker rejects anything outside `i16`.
 #[derive(Debug, Clone, Copy, PartialEq, Eq)]
 pub struct LevelModelFrameBoundsRecord {
     /// Model-local center in engine world units.
-    pub center: [i32; 3],
+    pub center: [i16; 3],
     /// Conservative sphere radius in engine world units.
-    pub radius: i32,
+    pub radius: i16,
 }
 
 /// One cooked PSX model: `.psxmdl` mesh, optional `.psxt`
